Send and recv

See merge request mallep/bomberman!3

+++ headers/client/client_requests.h
@@ -6,6 +6,6 @@
 
 int create_integrationrq(CReq *clientrq);
 int create_confrq(CReq *clientrq, game_mode_t game_mode, uint8_t id_player, uint8_t id_team);
-int create_ongamerq(CReq *clientrq, game_mode_t game_mode, uint8_t id_player, uint8_t id_team, uint32_t num, action_t action);
+int create_ongamerq(CReq *clientrq, game_mode_t game_mode, uint8_t id_player, uint8_t id_team, uint16_t num, action_t action);
 
 #endif
